// src/hooks/types.ts
export interface UserSubscription {
  user_id: string
  operational_project_id: string
  notifications: boolean
  reports: boolean
}

export interface FeedbackFormData {
  subject: string
  url: string
  comment: string
}

export interface Point {
  type: string
  coordinates: number[]
}

export interface MultiPolygon {
  type: string
  coordinates: number[][][][]
}

export interface SensorType {
  sensor_type_id: number
  name_short: string
  name_long: string
  name_metric: string
  unit: string | null
}

export interface DataType {
  data_type_id: number
  name_short: string
}

export interface Device {
  device_id: number
  device_type_id: number
  parent_device_id: number | null
  logical: boolean
  name_short: string | null
  name_long: string | null
  capacity_dc: number | null
  capacity_ac: number | null
  point: Point | null
  polygon: MultiPolygon | null
  device_type?: DeviceType
  name_full?: string
  cec_pv_module_id?: number
  pv_module_id?: number
  device_id_path?: string
}
// ...
export interface Event {
  event_id: number
  event_type_id: number | null
  device_id: number
  device: Device
  device_name_full: string
  time_start: string
  time_end: string | null
  time_detected: string
  time_last_analyzed: string | null
  failure_mode_id: number | null
  failure_mode: FailureMode | null
  root_cause_id: number | null
  loss_total_financial: number
}

export interface EventDeviceInfo {
  unique_types: {
    device_type_id: number
    device_type_name: string
  }[]
  unique_devices: {
    device_id: number
    device_name_full: string
  }[]
}

export interface PaginatedEvent {
  event_id: number
  device_name_full: string
  time_start: string
  time_end: string
  loss_daily_power: number
  loss_today_power: number
  loss_total_power: number
  loss_daily_financial: number
  loss_today_financial: number
  loss_total_financial: number
  root_cause: string
}
// ...
export interface UptimeData {
  device_id: number
  device_type_id: number
  device_name_full: string
  downtime_hours: number
  downtime_percentage: number
  events: number
}

export interface KPIProjectTemplate {
  data: {
    timestamps: string[]
    values: (number | null)[]
  }
}

export interface KPIDevicesTemplate {
  data: {
    timestamps: string[]
    names: string[]
    values: (number | null)[][]
  }
}

export interface FailureMode {
  failure_mode_id: number
  device_type_id: number
  name_short: string
  name_long: string
}

export interface RootCause {
  root_cause_id: number
  device_type_id: number
  name_short: string
  name_long: string
  name_full?: string
}

export interface Report {
  filename: string
  data_pdf: string
}

export interface ReportInstance {
  project_id: string
  report_type_id: number
  report_instance_id?: number
  is_visible: boolean
  report_type?: {
    report_type_id: number
    name_short: string
    name_long: string
    doc_url: string
  }
}

export interface RowData {
  name: string
  email: string
  company: string
}

export interface DataHeatmap {
  x: string[]
  y: string[]
  z: number[][]
}

export interface DataTimeSeries {
  x: string[]
  y: number[]
  y_range: number[]
  yaxis: string
  name: string
  sensor_type_name: string
  device_name_long: string
  tag_name_scada: string
  tag_name_long: string
  device_id: number
  sensor_type_id: number
}

export interface DegradationPOA {
  data: DataTimeSeries[]
  valid_indexes: string[]
  valid_columns: string[]
}
// ...
export interface SettlementPoint {
  settlement_point_id: number
  name: string
  settlement_point_type_id: number
  load_zone_id: number
  trading_hub_id: number
}

export interface QSE {
  qse_id: number
  name_short: string
  name_long: string
}

export interface DME {
  dme_id: number
  name_short: string
  name_long: string
}

export interface Resource {
  resource_id: number
  name_gen: string
  name_load: string
  name_long: string
  county: string
  in_service: number
  capacity_power: number
  qse_id: number
  dme_id: number
  settlement_point_id: number
  qse: QSE | null
  dme: DME | null
  settlement_point: SettlementPoint | null
}

export interface DatetimeDataFrame {
  index: string[]
  columns: string[]
  data: number[][]
}

export interface WeatherResponse {
  coord: {
    lon: number
    lat: number
  }
  weather: Array<{
    id: number
    main: string
    description: string
    icon: string
  }>
  base: string
  main: {
    temp: number
    feels_like: number
    temp_min: number
    temp_max: number
    pressure: number
    humidity: number
    sea_level: number
    grnd_level: number
  }
  visibility: number
  wind: {
    speed: number
    deg: number
    gust: number
  }
  clouds: {
    all: number
  }
  dt: number
  sys: {
    sunrise: number
    sunset: number
  }
  timezone: number
  id: number
  name: string
  cod: number
}

export interface ForecastResponse {
  cod: string
  message: number
  cnt: number
  list: {
    dt: number
    main: {
      temp: number
      feels_like: number
      temp_min: number
      temp_max: number
      pressure: number
      sea_level: number
      grnd_level: number
      humidity: number
      temp_kf: number
    }
    weather: {
      id: number
      main: string
      description: string
      icon: string
    }[]
    clouds: {
      all: number
    }
    wind: {
      speed: number
      deg: number
      gust: number
    }
    visibility: number
    pop: number
    sys: {
      pod: string
    }
    dt_txt: string
  }[]
  city: {
    id: number
    name: string
    coord: {
      lat: number
      lon: number
    }
    country: string
    population: number
    timezone: number
    sunrise: number
    sunset: number
  }
}
// ...
export interface SunburstProps {
  labels: string[]
  parents: string[]
  colors: string[]
  device_names: Record<string, number>
  hierarchy: Record<number, number[]>
}